code refactoring / added formatiing to output files

+++ no_voice_bot.py
@@ -1,24 +1,37 @@
+import os
+
+import ffmpeg
 from telegram.ext import Updater, MessageHandler, Filters
 from wit import Wit
-import ffmpeg
-import os
+
 
 CLIENT = Wit('CN76OWR2WNWJPNRKRRPUOTDZIHHCQ6L3')
 
+
 def echo(update, context):
-    """Echo the user message."""
+    """Extract user voice message and convert it to wav for subsequent post request to wit.ai."""
     audio_oga = context.bot.get_file(update.message.voice).download()
-    stream = ffmpeg.input(audio_oga)
-    stream = ffmpeg.output(stream, 'audio.wav')
-    ffmpeg.run(stream)
-    if os.path.exists(audio_oga):
+    filepath = '{0}_{1}.wav'.format(update.message.from_user['first_name'],
+                                    update.message.message_id)
+    # Run ffmpeg
+    (
+     ffmpeg
+     .input(audio_oga)
+     .output('{0}'.format(filepath))
+     .run()
+    )
+
+    with open(filepath, 'rb') as audio:
+        text = CLIENT.post_speech(audio)
+
+    voice_text = text['_text']
+    update.message.reply_text(voice_text, reply_to_message_id=update.message.message_id)
+
+    # Delete temporary files
+    if os.path.exists(audio_oga) and os.path.exists(filepath):
         os.remove(audio_oga)
-    audio = open('audio.wav', 'rb')
-    text = CLIENT.post_speech(audio)
-    if os.path.exists("audio.wav"):
-        os.remove("audio.wav")
-    final_text_json = text['_text']
-    update.message.reply_text(final_text_json, reply_to_message_id = update.message.message_id)
+        os.remove(filepath)
+
 
 def main():
     """Start the bot."""
@@ -29,6 +42,7 @@ def main():
     updater.start_polling()
     # Wait for other requests
     updater.idle()
+
 
 if __name__ == '__main__':
     main()
